requirements.txt
asgiref==3.8.1
backports.tarfile==1.2.0
bcrypt==4.1.2
build==1.2.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
check-manifest==0.50
cryptography==43.0.3
Django==5.1.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
docutils==0.21.2
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.5.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
packaging==24.2
pillow==10.2.0
pkginfo==1.10.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.9.0
pyotp==2.9.0
pyproject_hooks==1.2.0
python-decouple==3.8
pytz==2024.1
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
SecretStorage==3.3.3
sqlparse==0.4.4
twine==5.1.1
urllib3==2.2.3
zipp==3.21.0
